Fix codestyle

+++ src/main/java/ru/vk/itmo/proninvalentin/Metadata.java
@@ -1,10 +1,10 @@
 package ru.vk.itmo.proninvalentin;
 
 public class Metadata {
-    public static final long ENTRY_OFFSET_SIZE = Long.BYTES;
-    public static final long IS_DELETED_SIZE = Byte.BYTES;
-    public static final long CREATED_AT_SIZE = Long.BYTES;
-    public static final long SIZE = ENTRY_OFFSET_SIZE + IS_DELETED_SIZE + CREATED_AT_SIZE;
+    public static final byte ENTRY_OFFSET_SIZE = Long.BYTES;
+    public static final byte IS_DELETED_SIZE = Byte.BYTES;
+    public static final byte CREATED_AT_SIZE = Long.BYTES;
+    public static final short SIZE = ENTRY_OFFSET_SIZE + IS_DELETED_SIZE + CREATED_AT_SIZE;
     public final long entryOffset;
     public final boolean isDeleted;
     public final long createdAt;
